feat: simplify ponder config, added price discovery, uniswap v3 pool, optional chainlink oracle indexing
